organize and update booklist
