+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
 import Home from "./Containers/Home/Home";
 import BookCard from "./Components/BookCard/Bookcard";
@@ -17,80 +17,15 @@ import Search from "./Components/Search/search";
 import Detail from "./Components/Search/detail";
 import store from "../src/Components/books/store";
 import { Provider } from "react-redux";
-import axios from "axios";
-import { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import * as client from "../src/Components/books/client";
+
 function App() {
-  const [books, setBooks] = useState([]);
-  const API_BASE = "http://localhost:56100/api";
-  const URL = `${API_BASE}/books`;
-
-  const findAllBooks = async () => {
-    try {
-      const response = await client.findAllBooks();
-      setBooks(response);
-    } catch (error) {
-      console.error("Error fetching books:", error);
-    }
-  };
-
-  useEffect(() => {
-    findAllBooks();
-  }, []);
-  const [book, setBook] = React.useState({
-    name: "New book",
-    description: "book description",
-    author: "book author",
-  });
-
-  const addBook = async () => {
-    const response = await axios.post(URL, book);
-    setBooks([...books, response.data]);
-  };
-  const deleteBook = async (bookId) => {
-    try {
-      await axios.delete(`${URL}/${bookId}`);
-      setBooks((prevBooks) => prevBooks.filter((c) => c._id !== bookId));
-    } catch (error) {
-      console.error("Error deleting book:", error);
-    }
-  };
-  
-
-  const updateBook = async () => {
-    const response = await axios.put(`${URL}/${book._id}`, book);
-
-    setBooks(
-      books.map((c) => {
-        if (c._id === book._id) {
-          return book;
-        } else {
-          return c;
-        }
-      })
-    );
-  };
-  
   return (
     <Provider store={store}>
       <Router>
         <div className="App">
           <NavBar />
           <Routes>
-            <Route
-              path="/"
-              element={
-                <Home
-                  book={book}
-                  books={books}
-                  setBook={setBook}
-                  addBook={addBook}
-                  deleteBook={deleteBook}
-                  updateBook={updateBook}
-                />
-              }
-            />
+            <Route path="/" element={<Home />} />
             <Route path="/book" element={<BookCard />} />
             <Route path="/authorzone" element={<AuthorZone />} />
             <Route path="/myworks" element={<Myworks />} />
